requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-environ==0.12.0
django-filter==25.1
django-modeltranslation==0.19.12
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-rosetta==0.10.1
django-solo==2.4.0
djangorestframework==3.15.2
drf-yasg==1.21.9
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==8.13.55
pillow==11.1.0
polib==1.2.0
psycopg==3.2.5
psycopg-binary==3.2.5
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sorl-thumbnail==12.11.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
